test: RunnerPost 레포지터리 상세 조회 및 삭제 테스트 분리
