--- backend/baton/src/test/java/touch/baton/domain/runnerpost/repository/RunnerPostRepositoryTest.java
@@ -1,136 +1,0 @@
-package touch.baton.domain.runnerpost.repository;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import touch.baton.config.RepositoryTestConfig;
-import touch.baton.domain.common.vo.*;
-import touch.baton.domain.member.Member;
-import touch.baton.domain.member.repository.MemberRepository;
-import touch.baton.domain.member.vo.*;
-import touch.baton.domain.runner.Runner;
-import touch.baton.domain.runner.repository.RunnerRepository;
-import touch.baton.domain.runnerpost.RunnerPost;
-import touch.baton.domain.runnerpost.vo.Deadline;
-import touch.baton.domain.runnerpost.vo.PullRequestUrl;
-import touch.baton.domain.tag.RunnerPostTags;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Optional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
-class RunnerPostRepositoryTest extends RepositoryTestConfig {
-
-    @Autowired
-    private MemberRepository memberRepository;
-
-    @Autowired
-    private RunnerRepository runnerRepository;
-
-    @Autowired
-    private RunnerPostRepository runnerPostRepository;
-
-    @DisplayName("RunnerPost 식별자값으로 RunnerPost 정보와 Member 정보, Tag 정보를 조인하여 조회한다.")
-    @Test
-    void success_joinMemberProfileAndTagByRunnerPostId() {
-        // given
-        final Member member = Member.builder()
-                .memberName(new MemberName("헤에디주"))
-                .email(new Email("test@test.co.kr"))
-                .oauthId(new OauthId("dsigjh98gh230gn2oinv913bcuo23nqovbvu93b12voi3bc31j"))
-                .githubUrl(new GithubUrl("github.com/hyena0608"))
-                .company(new Company("우아한형제들"))
-                .build();
-        final Member saveMember = memberRepository.saveAndFlush(member);
-
-        final Runner runner = Runner.builder()
-                .totalRating(new TotalRating(100))
-                .grade(Grade.BARE_FOOT)
-                .member(saveMember)
-                .build();
-        final Runner saveRunner = runnerRepository.saveAndFlush(runner);
-
-        final LocalDateTime deadline = LocalDateTime.now();
-        final RunnerPost runnerPost = RunnerPost.builder()
-                .title(new Title("제 코드 리뷰 좀 해주세요!!"))
-                .contents(new Contents("제 코드는 클린코드가 맞을까요?"))
-                .deadline(new Deadline(deadline))
-                .pullRequestUrl(new PullRequestUrl("https://"))
-                .watchedCount(new WatchedCount(1))
-                .chattingRoomCount(new ChattingRoomCount(1))
-                .runnerPostTags(new RunnerPostTags(new ArrayList<>()))
-                .runner(saveRunner)
-                .supporter(null)
-                .build();
-        final Long saveRunnerPostId = runnerPostRepository.saveAndFlush(runnerPost).getId();
-
-        // when
-        final Optional<RunnerPost> maybeRunnerPost = runnerPostRepository.joinMemberByRunnerPostId(saveRunnerPostId);
-
-        // then
-        assertAll(
-                () -> assertThat(maybeRunnerPost).isPresent(),
-                () -> assertThat(maybeRunnerPost.get())
-                .usingRecursiveComparison()
-                .ignoringExpectedNullFields()
-                .isEqualTo(
-                RunnerPost.builder()
-                        .title(new Title("제 코드 리뷰 좀 해주세요!!"))
-                        .contents(new Contents("제 코드는 클린코드가 맞을까요?"))
-                        .deadline(new Deadline(deadline))
-                        .pullRequestUrl(new PullRequestUrl("https://"))
-                        .watchedCount(new WatchedCount(1))
-                        .chattingRoomCount(new ChattingRoomCount(1))
-                        .runnerPostTags(new RunnerPostTags(new ArrayList<>()))
-                        .runner(runner)
-                        .supporter(null)
-                        .build()
-                )
-        );
-    }
-
-    @DisplayName("RunnerPost 식별자값으로 RunnerPost 을 삭제한다.")
-    @Test
-    void success_deleteByRunnerPostId() {
-        // given
-        final Member member = Member.builder()
-                .memberName(new MemberName("헤에디주"))
-                .email(new Email("test@test.co.kr"))
-                .oauthId(new OauthId("dsigjh98gh230gn2oinv913bcuo23nqovbvu93b12voi3bc31j"))
-                .githubUrl(new GithubUrl("github.com/hyena0608"))
-                .company(new Company("우아한형제들"))
-                .build();
-        final Member saveMember = memberRepository.saveAndFlush(member);
-
-        final Runner runner = Runner.builder()
-                .totalRating(new TotalRating(100))
-                .grade(Grade.BARE_FOOT)
-                .member(saveMember)
-                .build();
-        final Runner saveRunner = runnerRepository.saveAndFlush(runner);
-
-        final RunnerPost runnerPost = RunnerPost.builder()
-                .title(new Title("제 코드 리뷰 좀 해주세요!!"))
-                .contents(new Contents("제 코드는 클린코드가 맞을까요?"))
-                .deadline(new Deadline(LocalDateTime.now()))
-                .pullRequestUrl(new PullRequestUrl("https://"))
-                .watchedCount(new WatchedCount(1))
-                .chattingRoomCount(new ChattingRoomCount(1))
-                .runnerPostTags(new RunnerPostTags(new ArrayList<>()))
-                .runner(saveRunner)
-                .supporter(null)
-                .build();
-        final Long saveRunnerPostId = runnerPostRepository.saveAndFlush(runnerPost).getId();
-
-        // when
-        runnerPostRepository.deleteById(saveRunnerPostId);
-
-        final Optional<RunnerPost> maybeRunnerPost = runnerPostRepository.findById(saveRunnerPostId);
-
-        // then
-        assertThat(maybeRunnerPost).isNotPresent();
-    }
-}
